[feat] Completes backend
Database schema added and seed file created, able to successfully
seed database. Sets up route to search endpoint and tested
with postman the ability to get results where the name is like
the input in our req.query. Next steps are to build tests and
begin working on the front end.